% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TopDecileLift.R
\name{TopDecileLift}
\alias{TopDecileLift}
\title{Compute the top decile lift and plot the lift curve}
\usage{
TopDecileLift(predicted, labels)
}
\arguments{
\item{predicted}{A numeric vector with the classifier's predicted scores / probabilities}

\item{labels}{An integer vector containing binary labels with values {0,1}}
}
\value{
a scalar denoting the top decile lift
}
\description{
\code{TopDecileLift} computes the commonly used top decile lift by ordering the data by the predictions, and computing the proportion of positives in the top 10\%.
}
\examples{
data(churn)
TopDecileLift(churn$predictions,churn$labels)
}
\author{
Steven Hoornaert, Michel Ballings, Dirk Van den Poel, Maintainer: \email{Steven.Hoornaert@UGent.be}
}

